crm_ingest/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crm_ingest"
version = "0.1.0"
description = "Local CRM ingestion agent for post-meeting voice memos"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.30.0",          # Whisper transcription
    "gspread>=6.0.0",          # Google Sheets
    "google-auth>=2.29.0",
    "pydantic>=2.7.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
crm_ingest = "crm_ingest.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["crm_ingest*"]
